api/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
# torch and torchvision installed manually for cpu
opencv-python-headless
requests
ultralytics
